feat(storage): supports storage

+++ admin3-server/src/main/java/tech/wetech/admin3/infra/storage/StorageServiceImpl.java
@@ -4,14 +4,17 @@ import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import tech.wetech.admin3.common.StringUtils;
-import tech.wetech.admin3.sys.model.storage.StorageConfig;
-import tech.wetech.admin3.sys.model.storage.StorageFile;
+import tech.wetech.admin3.sys.exception.StorageException;
+import tech.wetech.admin3.sys.model.StorageConfig;
+import tech.wetech.admin3.sys.model.StorageFile;
 import tech.wetech.admin3.sys.repository.StorageConfigRepository;
 import tech.wetech.admin3.sys.repository.StorageFileRepository;
 import tech.wetech.admin3.sys.service.StorageService;
 
 import java.io.InputStream;
 import java.util.List;
+
+import static tech.wetech.admin3.common.CommonResultStatus.FAIL;
 
 /**
  * @author cjbi
@@ -33,20 +36,62 @@ public class StorageServiceImpl implements StorageService {
   }
 
   @Override
-  public StorageConfig createConfig(StorageConfig storageConfig) {
-    storageConfigRepository.save(storageConfig);
-    return storageConfigRepository.save(storageConfig);
+  public StorageConfig getConfig(Long id) {
+    return storageConfigRepository.findById(id)
+      .orElseThrow(() -> new StorageException(FAIL, "存储配置不存在"));
   }
 
   @Override
-  public StorageConfig updateConfig(StorageConfig storageConfig) {
-    return storageConfigRepository.save(storageConfig);
+  @Transactional
+  public StorageConfig createConfig(String name, StorageConfig.Type type, String accessKey, String endpoint, String bucketName, String address, String storagePath) {
+    return storageConfigRepository.save(buildConfig(null, name, type, accessKey, endpoint, bucketName, address));
+  }
+
+  private StorageConfig buildConfig(Long id, String name, StorageConfig.Type type, String accessKey, String endpoint, String bucketName, String address) {
+    StorageConfig storageConfig = new StorageConfig();
+    storageConfig.setId(id);
+    storageConfig.setName(name);
+    storageConfig.setIsDefault(false);
+    storageConfig.setType(type);
+    storageConfig.setAccessKey(name);
+    storageConfig.setSecretKey(accessKey);
+    storageConfig.setEndpoint(endpoint);
+    storageConfig.setBucketName(bucketName);
+    storageConfig.setAddress(address);
+    storageConfig.setStoragePath(endpoint);
+    return storageConfig;
   }
 
   @Override
+  @Transactional
+  public StorageConfig updateConfig(Long id, String name, StorageConfig.Type type, String accessKey, String endpoint, String bucketName, String address, String storagePath) {
+    return storageConfigRepository.save(buildConfig(id, name, type, accessKey, endpoint, bucketName, address));
+  }
+
+  @Override
+  @Transactional
   public void deleteConfig(StorageConfig storageConfig) {
+    if (storageConfig.isDefault()) {
+      throw new StorageException(FAIL, "不能删除默认配置");
+    }
     storageConfigRepository.delete(storageConfig);
   }
+
+  @Override
+  @Transactional
+  public void markAsDefault(StorageConfig storageConfig) {
+    storageConfig.setIsDefault(true);
+    List<StorageConfig> configList = findConfigList();
+    for (StorageConfig record : configList) {
+      if (record.equals(storageConfig)) {
+        record.setIsDefault(true);
+      } else {
+        record.setIsDefault(false);
+      }
+    }
+    storageConfigRepository.saveAll(configList);
+  }
+
 
   @Override
   @Transactional
@@ -74,9 +119,9 @@ public class StorageServiceImpl implements StorageService {
     return new S3Storage(config);
   }
 
-  private String generateKey(String originalFilename) {
-    int index = originalFilename.lastIndexOf('.');
-    String suffix = originalFilename.substring(index);
+  private String generateKey(String filename) {
+    int index = filename.lastIndexOf('.');
+    String suffix = filename.substring(index);
 
     String key = null;
     StorageFile storageFile = null;
@@ -89,8 +134,9 @@ public class StorageServiceImpl implements StorageService {
   }
 
   @Override
-  public void delete(String keyName) {
-    getStorage().delete(keyName);
+  public void delete(String key) {
+    getStorage().delete(key);
+    storageFileRepository.deleteByKey(key);
   }
 
   @Override
